Separate Deck Creation form and styles from DecklistScreen
