+++ styles/decks/DecklistScreenStyle.ts
@@ -1,0 +1,17 @@
+import { StyleSheet } from "react-native";
+
+export const DecklistScreenStyle = StyleSheet.create({
+  container: {
+    alignItems: "center",
+    justifyContent: "center",
+    padding: 12,
+    paddingTop: 64,
+  },
+  decksList: {
+    minWidth: "80%",
+  },
+  title: {
+    fontSize: 24,
+    textAlign: "center",
+  },
+});
